fix(format-pack): three audit-follow-up corrections in new walkers
Audit (with independent verification) flagged three silent-wrong-value
bugs in the format-pack walkers landed in 12eeb57. Each fix carries a
test that fails against the pre-fix code and passes after.

1. MPC SV8 EI packet — encoder-version byte ordering. The C reference
   (3rdparty/musepack_src_r475/libmpcdec/streaminfo.c:228-230) shifts
   the three bytes by <<24, <<16, <<8 (low byte 0). The C# port shifted
   by <<16, <<8 and used the un-shifted final byte directly, so 1.30.0
   surfaced as 0x0001_1E00 instead of 0x011E_0000.

2. WavPack sub-block UniqueId mask — was 0x1F, must be 0x3F. Per
   3rdparty/WavPack/include/wavpack.h:160-163 only ID_LARGE (0x80) and
   ID_ODD_SIZE (0x40) are flag bits; ID_OPTIONAL_DATA (0x20) is part of
   the unique-id space (e.g., ID_RIFF_HEADER == ID_OPTIONAL_DATA | 0x01
   == 0x21). The 0x1F mask silently turned 0x21 into 0x01.

3. MAC seek-table 32-bit wrap correction. Files larger than 4 GiB store
   wrapped 32-bit offsets that the upstream C++ reference (3rdparty/
   MAC_1284_SDK/Source/MACLib/APEHeader.cpp:116-131,
   Convert32BitSeekTable) corrects by adding 0x1_0000_0000 each time a
   value decreases relative to the previous one. The C# port stored the
   raw uint and silently rolled over. Widens MacSeekEntry.FileOffset
   from uint to long; the BuildFrameTable arithmetic already widens
   through MacFrame's long fields and MacDescriptor.TotalApeFrameDataBytes.

+++ AudioVideoLib/Formats/MacSeekEntry.cs
@@ -6,9 +6,16 @@ namespace AudioVideoLib.Formats;
 /// <see cref="MacDescriptor.SeekTableBytes"/> / 4. See
 /// <c>3rdparty/MAC_1284_SDK/Source/MACLib/APEHeader.cpp:268</c>.
 /// </summary>
+/// <remarks>
+/// The on-disk offsets are 32-bit, but for files larger than 4 GiB the upstream encoder
+/// allows them to wrap. <see cref="MacStream"/> applies the same wrap-correction as
+/// <c>Convert32BitSeekTable</c> (<c>APEHeader.cpp:116-131</c>) — accumulating
+/// <c>0x1_0000_0000</c> each time the raw value decreases — so <see cref="FileOffset"/>
+/// is widened to <see cref="long"/> and represents the corrected absolute offset.
+/// </remarks>
 public sealed class MacSeekEntry
 {
-    internal MacSeekEntry(int frameIndex, uint fileOffset)
+    internal MacSeekEntry(int frameIndex, long fileOffset)
     {
         FrameIndex = frameIndex;
         FileOffset = fileOffset;
@@ -17,6 +24,6 @@ public sealed class MacSeekEntry
     /// <summary>Gets the index of the frame this entry refers to (0-based).</summary>
     public int FrameIndex { get; }
 
-    /// <summary>Gets the absolute file offset of the frame's first byte.</summary>
-    public uint FileOffset { get; }
+    /// <summary>Gets the absolute file offset of the frame's first byte (post wrap-correction).</summary>
+    public long FileOffset { get; }
 }
